Inventory: RMB to drop item to Ground in preBattleEquip screen
- if not grenade
- inspired by Hobbes; still should do RMB-to-equip, same as Ctrl+LMB.

+++ src/Battlescape/ActionMenuState.h
@@ -50,7 +50,7 @@ private:
 	void addItem(
 			BattleActionType baType,
 			const std::string& desc,
-			int* id);
+			size_t* id);
 
 
 	public:
